backend/requirements.txt
# FastAPI core
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.29
alembic==1.13.1
aiosqlite==0.20.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI / NLP (OpenAI)
openai==1.30.1
faiss-cpu==1.8.0
numpy==1.26.4

# Text processing
tiktoken==0.7.0
nltk==3.8.1
textblob==0.18.0

# Export
reportlab==4.1.0
pandas==2.2.2

# Config / utils
pydantic-settings==2.2.1
pydantic[email]==2.7.1
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
